requirements.txt
amqp==5.0.9
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.3
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.11
djangorestframework==3.13.1
importlib-metadata==4.10.0
kombu==5.2.3
prompt-toolkit==3.0.24
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
typing_extensions==4.0.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
zipp==3.7.0

pip~=21.3.1
wheel~=0.37.0
setuptools~=59.6.0